Fixed bug where arrays of objects would get infinitely extended.
When you create a new instance of an object based on AjaxSolr.Class,
and set one of the parameters to an array of objects, this code would
iterate through each one of those objects and extend it.

This fix avoids calling extend if the array is actually an array
object.

Tested and works with gh-pages, and my own code that was previously
broken because of this bug.